requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.0.1
coverage==7.3.1
crispy-bootstrap5==0.7
cryptography==39.0.1
defusedxml==0.7.1
dependency-checker==0.0.5
Deprecated==1.2.14
distlib==0.3.7
dj-rest-auth==5.0.1
Django==4.2
django-allauth==0.52.0
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-filter==23.3
django-js-asset==2.1.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
drf-spectacular==0.26.5
exceptiongroup==1.1.3
Faker==19.6.2
filelock==3.12.4
gunicorn==20.1.0
identify==2.5.30
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.1
pipdeptree==2.13.0
pipreqs==0.4.13
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.5.0
psycopg==3.1.8
psycopg-binary==3.1.8
pycparser==2.21
PyGithub==2.1.1
PyJWT==2.6.0
pymemcache==4.0.0
PyNaCl==1.5.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.8
python-memcached==1.59
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.30.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rpds-py==0.10.3
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.24.5
whitenoise==6.4.0
wrapt==1.15.0
yarg==0.1.9
